+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Deprecate NullableParamStyleFixer
 - Deprecate SingleLineThrowFixer
 - Allow symfony/finder 5.0
+- Add Windows OS support with AppVeyor
 
 ## v1.16.0 - *2019-10-24*
 - Add PhpdocOnlyAllowedAnnotationsFixer